requirements.txt
aiogram==2.11.2
aiohttp==3.7.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Django==3.1.4
djangorestframework==3.12.2
dlib==19.21.1
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==2.10
multidict==5.1.0
numpy==1.19.4
Pillow==8.0.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
yarl==1.6.3
django-filter==2.4.0
djoser
psycopg2-binary

